Add fairly fine grained filtering support
Now the program can make only one query but it can accept multiple
optional filters based on the fields. Each extra argument is a filter
specification, and filters are all joined using a logical AND.

The individual filters have the format X:fitler, there X is a field
specification:
t = titulo
d = desc
a = autor
f = formato
c = comentarios
C = cds
F = fecha
D = downloads

And filter is a string that should be found on that field (case
insensitive). If the format specifier is not known (or there isn't one)
the filter string is looked in all the fields.